Moved NetworkServiceHelper into its own file.
